pyproject.toml
[tool.poetry]
name = "llamaindex-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
streamlit = "^1.31.1"
openai = "^1.12.0"
chromadb = "^0.4.23"
llama-index = "^0.10.12"
traceloop-sdk = "^0.12.5"
llama-index-postprocessor-cohere-rerank = "^0.1.2"
sqlalchemy = "^2.0.27"
wikipedia = "^1.4.0"
llama-index-readers-wikipedia = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
